open_meteo_api/requirements.txt
openmeteo-requests==1.1.0
pandas==2.2.2
requests-cache==1.2.0
retry-requests==1.0.0
mysql-connector-python==8.4.0
